requirements.txt
google-generativeai
rich
streamlit
pandas
automata-lib
graphviz
